Go: Update to Go 1.24.6 (#697)
