+++ interpreter/golabels/integrationtests/.gitignore
@@ -1,0 +1,1 @@
+pprof_1_*
