+++ interpreter/golabels/integrationtests/pprof/go.mod
@@ -1,0 +1,3 @@
+module go.opentelemetry.io/ebpf-profiler/interpreter/golabels/integrationtests/pprof
+
+go 1.23.7
